Can that dialog be like an expanded dialog off the nav item similar to a
