+++ src/app/layout.tsx
@@ -22,7 +22,6 @@ import { NewSprintDialog } from '@/components/new-sprint-dialog';
 import { EditSprintDialog } from '@/components/edit-sprint-dialog';
 import { EditTaskDialog } from '@/components/edit-task-dialog';
 import { NewUserStoryDialog } from '@/components/new-user-story-dialog';
-import { StarredItemsDialog } from '@/components/starred-items-dialog';
 import { useIdle } from '@/hooks/use-idle';
 import { signOut } from '@/services/auth-service';
 import { useToast } from '@/hooks/use-toast';
@@ -127,7 +126,6 @@ export default function RootLayout({
                   <EditSprintDialog />
                   <EditTaskDialog />
                   <NewUserStoryDialog />
-                  <StarredItemsDialog />
                   <AssessmentModal 
                     // These props are managed by the QuickActionProvider now
                   />
