+++ src/components/sidebar.tsx
@@ -2,7 +2,7 @@
 'use client';
 
 import * as React from 'react';
-import { usePathname } from 'next/navigation';
+import { usePathname, useRouter } from 'next/navigation';
 import Link from 'next/link';
 import {
   ChevronsLeft,
@@ -16,6 +16,7 @@ import {
   Library,
   BookCopy,
   Star,
+  Search,
 } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
@@ -27,6 +28,12 @@ import {
 } from '@/components/ui/tooltip';
 import { Separator } from './ui/separator';
 import { useQuickAction } from '@/contexts/quick-action-context';
+import { Popover, PopoverContent, PopoverTrigger } from './ui/popover';
+import { Input } from './ui/input';
+import { ScrollArea } from './ui/scroll-area';
+import { getProjects, Project } from '@/services/project-service';
+import { getAssessments, Assessment } from '@/services/assessment-service';
+import { Skeleton } from './ui/skeleton';
 
 const NavGroup = ({
     title,
@@ -50,9 +57,90 @@ const NavGroup = ({
     )
 }
 
+type StarredItem = (Project | Assessment) & { itemType: 'Engagement' | 'Assessment' };
+
+function StarredItemsPopoverContent() {
+    const [items, setItems] = React.useState<StarredItem[]>([]);
+    const [loading, setLoading] = React.useState(false);
+    const [searchTerm, setSearchTerm] = React.useState('');
+    const router = useRouter();
+
+    React.useEffect(() => {
+        const fetchStarredItems = async () => {
+            setLoading(true);
+            const [projects, assessments] = await Promise.all([getProjects(), getAssessments()]);
+            
+            const starredProjects = projects
+                .filter(p => p.isStarred)
+                .map(p => ({ ...p, itemType: 'Engagement' as const }));
+                
+            const starredAssessments = assessments
+                .filter(a => a.isStarred)
+                .map(a => ({ ...a, itemType: 'Assessment' as const }));
+
+            setItems([...starredProjects, ...starredAssessments].sort((a,b) => a.name.localeCompare(b.name)));
+            setLoading(false);
+        };
+        fetchStarredItems();
+    }, []);
+
+    const filteredItems = items.filter(item => 
+        item.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
+        item.itemType.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+    
+    const handleItemClick = (item: StarredItem) => {
+        if (item.itemType === 'Engagement') {
+            router.push(`/dashboard/projects/${item.id}`);
+        } else {
+            router.push(`/assessment/${item.id}/report`);
+        }
+    };
+
+    return (
+        <PopoverContent className="w-80" side="right" align="start">
+            <h4 className="font-medium text-sm mb-2">Starred Items</h4>
+            <div className="relative mb-4">
+                <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
+                <Input 
+                    placeholder="Search starred items..."
+                    className="pl-8"
+                    value={searchTerm}
+                    onChange={(e) => setSearchTerm(e.target.value)}
+                />
+            </div>
+            <ScrollArea className="h-72">
+                <div className="space-y-2 pr-4">
+                    {loading ? (
+                        Array.from({length: 3}).map((_, i) => <Skeleton key={i} className="h-12 w-full" />)
+                    ) : filteredItems.length > 0 ? (
+                        filteredItems.map(item => (
+                            <div 
+                                key={`${item.itemType}-${item.id}`} 
+                                className="flex items-center gap-3 p-2 rounded-md hover:bg-muted cursor-pointer"
+                                onClick={() => handleItemClick(item)}
+                            >
+                                {item.itemType === 'Engagement' ? 
+                                    <FolderKanban className="h-5 w-5 text-primary" /> : 
+                                    <ClipboardList className="h-5 w-5 text-primary" />
+                                }
+                                <div className="flex-1">
+                                    <p className="text-sm font-medium">{item.name}</p>
+                                    <p className="text-xs text-muted-foreground">{item.itemType}</p>
+                                </div>
+                            </div>
+                        ))
+                    ) : (
+                        <p className="text-sm text-muted-foreground text-center py-10">No starred items found.</p>
+                    )}
+                </div>
+            </ScrollArea>
+        </PopoverContent>
+    )
+}
+
 export function Sidebar() {
   const pathname = usePathname();
-  const { openStarredItemsDialog } = useQuickAction();
   const [isCollapsed, setIsCollapsed] = React.useState(false);
 
   const navItems = [
@@ -60,7 +148,7 @@ export function Sidebar() {
         group: 'HOME',
         links: [
             { href: `/dashboard`, label: 'Dashboard', icon: Home },
-            { action: openStarredItemsDialog, label: 'Starred', icon: Star }
+            { id: 'starred', label: 'Starred', icon: Star }
         ]
     },
     {
@@ -91,6 +179,74 @@ export function Sidebar() {
     setIsCollapsed(!isCollapsed);
   }
 
+  const renderNavItem = (item: any) => {
+    const Icon = item.icon;
+    const isActive = 'href' in item && pathname === item.href;
+
+    if (item.id === 'starred') {
+        return (
+             <Popover>
+                <Tooltip>
+                    <TooltipTrigger asChild>
+                        <PopoverTrigger asChild>
+                             <Button
+                                variant="sidebar"
+                                className='w-full justify-start relative'
+                                >
+                                <Icon
+                                className={cn('h-5 w-5', {
+                                    'mr-2': !isCollapsed,
+                                })}
+                                />
+                                {!isCollapsed && item.label}
+                            </Button>
+                        </PopoverTrigger>
+                    </TooltipTrigger>
+                    {isCollapsed && (
+                        <TooltipContent side="right">
+                        {item.label}
+                        </TooltipContent>
+                    )}
+                </Tooltip>
+                <StarredItemsPopoverContent />
+            </Popover>
+        )
+    }
+
+    return (
+        <Tooltip>
+            <TooltipTrigger asChild>
+                <Button
+                asChild
+                variant="sidebar"
+                className={cn(
+                    'w-full justify-start relative',
+                    isActive &&
+                    'bg-sidebar-accent text-sidebar-accent-foreground'
+                )}
+                >
+                    <Link href={item.href}>
+                        {isActive && (
+                        <div className="absolute left-0 top-0 h-full w-1.5 bg-primary" />
+                        )}
+                        <Icon
+                        className={cn('h-5 w-5', {
+                            'mr-2': !isCollapsed,
+                        })}
+                        />
+                        {!isCollapsed && item.label}
+                    </Link>
+                </Button>
+            </TooltipTrigger>
+            {isCollapsed && (
+                <TooltipContent side="right">
+                {item.label}
+                </TooltipContent>
+            )}
+        </Tooltip>
+    )
+  }
+
   return (
     <>
       <TooltipProvider delayDuration={0}>
@@ -110,55 +266,7 @@ export function Sidebar() {
                   if (group.group === 'HOME') {
                     return (
                         <div key={group.group}>
-                            {group.links.map((item) => {
-                                const Icon = item.icon;
-                                const isActive = 'href' in item && pathname === item.href;
-                                
-                                const buttonContent = 'href' in item ? (
-                                    <Link href={item.href}>
-                                        {isActive && (
-                                        <div className="absolute left-0 top-0 h-full w-1.5 bg-primary" />
-                                        )}
-                                        <Icon
-                                        className={cn('h-5 w-5', {
-                                            'mr-2': !isCollapsed,
-                                        })}
-                                        />
-                                        {!isCollapsed && item.label}
-                                    </Link>
-                                ) : (
-                                    <div className="flex items-center" onClick={item.action}>
-                                        <Icon
-                                        className={cn('h-5 w-5', {
-                                            'mr-2': !isCollapsed,
-                                        })}
-                                        />
-                                        {!isCollapsed && item.label}
-                                    </div>
-                                )
-                                return (
-                                <Tooltip key={item.label}>
-                                    <TooltipTrigger asChild>
-                                        <Button
-                                        asChild={'href' in item}
-                                        variant="sidebar"
-                                        className={cn(
-                                            'w-full justify-start relative',
-                                            isActive &&
-                                            'bg-sidebar-accent text-sidebar-accent-foreground'
-                                        )}
-                                        >
-                                        {buttonContent}
-                                        </Button>
-                                    </TooltipTrigger>
-                                    {isCollapsed && (
-                                        <TooltipContent side="right">
-                                        {item.label}
-                                        </TooltipContent>
-                                    )}
-                                    </Tooltip>
-                                )
-                            })}
+                            {group.links.map((item) => renderNavItem(item))}
                         </div>
                     )
                   }
@@ -179,8 +287,8 @@ export function Sidebar() {
                                     'bg-sidebar-accent text-sidebar-accent-foreground'
                                 )}
                               >
-                                {'href' in item ? (
-                                    <Link href={item.href}>
+                                
+                                    <Link href={'href' in item ? item.href : '#'}>
                                     {isActive && (
                                         <div className="absolute left-0 top-0 h-full w-1.5 bg-primary" />
                                     )}
@@ -191,16 +299,7 @@ export function Sidebar() {
                                     />
                                     {!isCollapsed && item.label}
                                     </Link>
-                                ) : (
-                                    <div className="flex items-center" onClick={'action' in item ? item.action : undefined}>
-                                        <Icon
-                                            className={cn('h-5 w-5', {
-                                            'mr-2': !isCollapsed,
-                                            })}
-                                        />
-                                        {!isCollapsed && item.label}
-                                    </div>
-                                )}
+                                
                               </Button>
                             </TooltipTrigger>
                             {isCollapsed && (
